add try/except in market order, add max fail, also only truncate table in other tasks if request have results

+++ lazyblacksmith/tasks/adjusted_price.py
@@ -16,25 +16,25 @@ import pytz
 @celery_app.task(name="schedule.update_adjusted_price")
 def update_adjusted_price():
     # delete everything from table first.
-    db.engine.execute("TRUNCATE TABLE %s" % ItemAdjustedPrice.__tablename__)
-    db.session.commit()
-
     item_adjusted_price = []
     count = 0
 
     market_price = esiclient.request(get_markets_prices())
-    for item_price in market_price.data:
-        count += 1
-        item_adjusted_price.append({
-            'item_id': item_price.type_id,
-            'price': item_price.adjusted_price or 0.00
-        })
+    
+    if market_price.status == 200:
+        for item_price in market_price.data:
+            count += 1
+            item_adjusted_price.append({
+                'item_id': item_price.type_id,
+                'price': item_price.adjusted_price or 0.00
+            })
 
-    db.engine.execute(
-        ItemAdjustedPrice.__table__.insert(),
-        item_adjusted_price
-    )
-    db.session.commit()
+        db.engine.execute("TRUNCATE TABLE %s" % ItemAdjustedPrice.__tablename__)
+        db.engine.execute(
+            ItemAdjustedPrice.__table__.insert(),
+            item_adjusted_price
+        )
+        db.session.commit()
     
     task_status = TaskStatus(
         name='schedule.update_adjusted_price',
