+++ lazyblacksmith/tasks/industry_index.py
@@ -16,30 +16,28 @@ import pytz
 @celery_app.task(name="schedule.update_industry_indexes")
 def update_industry_index():
     """ Get the industry indexes list from API. """
-    db.engine.execute("TRUNCATE TABLE %s" % IndustryIndex.__tablename__)
-    db.session.commit()
-
     all_indexes = esiclient.request(get_industry_systems())
-
     insert_index_list = []
 
-    for index in all_indexes.data:
-        solar_system = index.solar_system_id
+    if all_indexes.status == 200:
+        for index in all_indexes.data:
+            solar_system = index.solar_system_id
 
-        for activity_index in index.cost_indices:
-            row = {}
-            row['solarsystem_id'] = solar_system
-            row['activity'] = IndustryIndex.activity_string_to_activity(
-                activity_index.activity
-            )
-            row['cost_index'] = activity_index.cost_index
-            insert_index_list.append(row)
+            for activity_index in index.cost_indices:
+                row = {}
+                row['solarsystem_id'] = solar_system
+                row['activity'] = IndustryIndex.activity_string_to_activity(
+                    activity_index.activity
+                )
+                row['cost_index'] = activity_index.cost_index
+                insert_index_list.append(row)
 
-    db.engine.execute(
-        IndustryIndex.__table__.insert(),
-        insert_index_list
-    )
-    db.session.commit()
+        db.engine.execute("TRUNCATE TABLE %s" % IndustryIndex.__tablename__)
+        db.engine.execute(
+            IndustryIndex.__table__.insert(),
+            insert_index_list
+        )
+        db.session.commit()
 
     task_status = TaskStatus(
         name='schedule.update_industry_indexes',
